refactor: put controller in their own module
